Implemented changes s.t when GetRoomsByUser is implemneted YourEvents will work -Removed the dummy data from YourEvents and put it as return from GetRoomsByUser-Added more animations to YourEvents
