+++ src/components/GetRoomsByUser/FilterOldRooms.js
@@ -2,18 +2,21 @@ import db from "../Db/firebaseConnect";
 import moment from "moment";
 import {DATE_FORMAT} from "../../constants";
 
-
+/**
+ * Removes old rooms from the Db
+ */
 const removeRooms = async ({oldRooms,email})=>{
     const _removeRoom = (roomId) =>{
         db.ref().transaction(()=>{
-            // remove entry from the /users
-            db.ref('/users/'+email+'/active_rooms/'+roomId).remove();
-            // remove entry from the /rooms
-            db.ref('/rooms/'+roomId).remove()
+                // remove entry from the /users
+                db.ref('/users/'+email+'/active_rooms/'+roomId).remove();
+                // remove entry from the /rooms
+                db.ref('/rooms/'+roomId).remove()
             }
-        );
+        ).catch((e)=>console.error("Error in deleting a room",e));
     };
-    oldRooms.forEach((room)=>_removeRoom(room.id));
+
+    oldRooms.forEach((room)=>_removeRoom(room.roomId));
 };
 
 /**
@@ -21,11 +24,11 @@ const removeRooms = async ({oldRooms,email})=>{
  * and launches a async function in the background to delete the old
  * rooms
  * @param email - email address of the user in question
- * @param rooms - a list of rooms from dB each room needs to have
- *                a extra parameter called "id" which holds the "id"
- *                if you get data from Db this "id" is automatically exluced
- *                so must add this parameter manually, eg room["id"]=1, for
- *                each room. 
+ * @param rooms - a list of rooms from dB each room. Note each room needs to have
+ *                a extra parameter called "roomId" which holds the roomId.
+ *                If you get data from Db this "roomId" is automatically excluded
+ *                so must add this parameter manually, eg room["roomId"]=1, for
+ *                each room.
  * @return list of current rooms
  */
 const FilterOldRooms = ({email,rooms}) =>{
@@ -46,7 +49,7 @@ const FilterOldRooms = ({email,rooms}) =>{
     };
 
     let currentRooms = rooms.reduce(_filterRooms,[]);
-    // async remove of old rooms
+    // async remove of old rooms -- don't care about promise
     removeRooms({email,oldRooms});
     return currentRooms;
