+++ src/views/YourEvents/YourEvents.scss
@@ -48,6 +48,8 @@
 .yourevents__container-list {
   background-image: url("./Images/cream_dust.png");
   background-repeat: repeat;
+  animation: yourevents_fade-in 1500ms 500ms forwards;
+  animation-delay: 0ms;
 
 }
 
@@ -65,6 +67,8 @@
 
   margin-left: auto;
   margin-right: auto;
+  animation: yourevents_fade-in 1500ms 500ms forwards;
+  animation-delay: 0ms;
 }
 
 @media only screen and (max-width: $mobile-break) {
